chore: move BrowserDetector to detectors directory
