+++ packages/opentelemetry-resources/test/detectors/node/BrowserDetector.test.ts
@@ -1,0 +1,43 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import * as sinon from 'sinon';
+import { Resource } from '../../../src';
+import { browserDetector } from '../../../src/detectors/BrowserDetector';
+import {
+  assertEmptyResource,
+} from '../../util/resource-assertions';
+
+
+describe('browserDetector()', () => {
+  beforeEach(() => {
+    (globalThis.window as {}) = {};
+    sinon.stub(globalThis, 'window').value({
+      navigator: {
+        userAgent: '',
+      }
+    });
+  });
+
+  afterEach(() => {
+    sinon.restore();
+  });
+
+  it('should return empty resources if window.document is missing', async () => {
+    const resource: Resource = await browserDetector.detect();
+    assertEmptyResource(resource);
+  });
+});
+
